+++ game/configuration.py
@@ -1,6 +1,22 @@
 # Colors from https://flatuicolors.com/palette/defo
 
-from game.custom import *
+from typing import Tuple
+
+
+class Pixel(int):
+    pass
+
+
+class RGB(Tuple[int, int, int, int | None]):
+    pass
+
+
+class RGB01(Tuple[int, int, int, int | None]):
+    pass
+
+
+class HEX(str):
+    pass
 
 
 class CornerMargin:
